refactor(node): move main to its own file and keep run infrastructure reusable
