+++ crates/papyrus_node/src/run_test.rs
@@ -1,12 +1,17 @@
 use std::time::Duration;
 
 use metrics_exporter_prometheus::PrometheusBuilder;
-use papyrus_node::config::NodeConfig;
 use papyrus_storage::{open_storage, StorageConfig};
 use papyrus_test_utils::prometheus_is_contained;
 use tempfile::TempDir;
 
-use crate::{run_threads, spawn_storage_metrics_collector, PapyrusResources, PapyrusTaskHandles};
+use crate::config::NodeConfig;
+use crate::run::{
+    run_threads,
+    spawn_storage_metrics_collector,
+    PapyrusResources,
+    PapyrusTaskHandles,
+};
 
 // The mission of this test is to ensure that if an error is returned from one of the spawned tasks,
 // the node will stop, and this error will be returned. This is done by checking the case of an
